fix for real bluegene systems to remove warnings

+++ trunk/src/smap/configure_functions.c
@@ -573,7 +573,7 @@ static int _resolve(char *com)
 #ifdef HAVE_BG_FILES
 	int len=strlen(com);
 	char *rack_mid = NULL;
-	int *coord = NULL;
+	uint16_t *coord = NULL;
 #endif
 
 	while(com[i] != '\0') {
